+++ src/alecto-external/alecto-pako-wrapper/alecto-pako-wrapper-native.js
@@ -1,9 +1,8 @@
-import Pako from "pako";
 
 export function alectoPakoGzip(s){
-    return Pako.gzip(s);
+    return ""
 }
 
 export function alectoPakoUnGzip(s){
-    return Pako.ungzip(s)
+    return ""
 }