+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 
 
+
+
 #### v0.2a
 
 - Refactor the code using TypeScript
@@ -12,4 +14,14 @@
 - Revert traits in `0.1e`
   - Remove supports for platform `tmall`
   - Remove supports for comment snapshot
+
+
+
+
+
+#### v0.2b
+
+- Supports platform `tmall` (Fix problem in `v0.1e`)
+- Supports snapshot for `taobao`
+- Fix bug: Incorrect multimedia comment identification